Improve: Query Playing Media with API
