+++ src/mudlet.pro
@@ -660,6 +660,7 @@ SOURCES += \
     TLuaInterpreter.cpp \
     TLuaInterpreterDiscord.cpp \
     TLuaInterpreterMapper.cpp \
+    TLuaInterpreterMedia.cpp \
     TLuaInterpreterMudletObjects.cpp \
     TLuaInterpreterNetworking.cpp \
     TLuaInterpreterUI.cpp \
